Desenvolvimento

+++ src/domains/Artist/service/serviceArtist.ts
@@ -1,32 +1,27 @@
-import { PrismaClient } from '@prisma/client';
+import prisma from '../../../../config/client';
 import { Artist } from '../model/modelArtist';
+import { QueryError } from '../../../../errors/QueryError';
 
-const prisma = new PrismaClient();
 
 class serviceArtist {
 	async create(body: Artist) {
 		if(body.name == '') {
-			throw new Error('O artista precisa de um nome!');
+			throw new QueryError('O artista precisa de um nome!');
 		}
 
 		if(!isNaN(Number(body.photo)) || body.photo == '') {
-			throw new Error('A foto precisa ser um link!');
+			throw new QueryError('A foto precisa ser um link!');
 		}
 
 		if(isNaN(body.streams)) {
-			throw new Error('As streams são contadas em números!');
-		}
-
-		if(isNaN(body.artistID)) {
-			throw new Error('O ID do artista precisa ser um número!');
+			throw new QueryError('As streams são contadas em números!');
 		}
 
 		const newArtist = await prisma.artist.create({
 			data: {
-				id: body.artistID,
 				name: body.name,
 				photo: body.photo,
-				streams: body.streams,
+				streams: Number(body.streams),
 			},
 		});
 
@@ -40,39 +35,55 @@ class serviceArtist {
 
 	async update(body: Artist) {
 		if(body.name == '') {
-			throw new Error('O artista precisa de um nome!');
+			throw new QueryError('O artista precisa de um nome!');
 		}
 
 		if(!isNaN(Number(body.photo)) || body.photo == '') {
-			throw new Error('A foto precisa ser um link!');
+			throw new QueryError('A foto precisa ser um link!');
 		}
 
-		if(isNaN(body.streams)) {
-			throw new Error('As streams são contadas em números!');
+		if(isNaN(Number(body.streams))) {
+			throw new QueryError('As streams são contadas em números!');
 		}
 
-		if(isNaN(body.artistID)) {
-			throw new Error('O ID do artista precisa ser um número!');
+		if (!await this.findArtist(Number(body.artistID))){
+			throw new QueryError('O artista não existe');
 		}
 
 		const atualizar = await prisma.artist.update({
-			where: {id: body.artistID},
+			where: {id: Number(body.artistID)},
 			data: {
 				name: body.name,
 				photo: body.photo,
-				streams: body.streams,
+				streams: Number(body.streams),
 			},
 		});
 
 		return atualizar;
 	}
 
-	async delete(body: Artist) {
+	async delete(id: number) {
+		if(isNaN(Number(id))) {
+			throw new QueryError('ID inválido');
+		}
+		if (!await this.findArtist(Number(id))){
+			throw new QueryError('O artista não existe');
+		}
 		const deletar = await prisma.artist.delete({
-			where: {id: body.artistID},
-		});
+			where: {id: Number(id)},
+		}); 	
 
 		return deletar;
+	}
+
+	async findArtist(id: number){
+		if (id == 0 || isNaN(Number(id))){
+			throw new QueryError('O id do artista precisa ser um número');
+		}
+		const find = await prisma.artist.findUnique({
+			where:{id: Number(id)}
+		});
+		return find;
 	}
 }
